fix(engine): streamprocessor can recover on replay mode when no events available
When starting on replay mode, the events that are already applied in the snapshot may not be available in the logstreams. This can happen because after the snapshot is received, the replication of events after the snapshot index is delayed. In this case, we don't want to fail the recovery. ReplayStateMachine can wait and apply the events when they are available.