fix: reduce timeouts
Previously we have set the timeouts to 2 and 15 seconds, which are netty defaults. This caused to delay the closing of a broker by ~4 seconds, since the quite period in the netty service is waiting 2 seconds until it is really turning off. If there is a request coming in it will wait another 2 seconds. This can happen until the max timeout of 15 seconds is reached. In general we don't want to postpone the shutdown of the broker for so long, especially not in our integration tests. Reducing the timeouts to a minimum caused reducing the an example test case by ~ 7-8 seconds. We want to try out the low values first in benchmarks and production if we see no issues with that we will remove them completely and use the normal shutdown.