fix: increase clock to trigger snapshot
Previously the snapshot time was not increased, which caused flaky tests. It happened depending on which was the previous leader and later the leader that maybe an snapshot was taken already previously, then this test succeeded. If the latest leader was at the begin follower it might happen that this node didn't took a snapshot before, this caused the test to fail.