fix(broker): do not fail exporter if position not found when recovering
On followers it is posssible that the log doesnot contain the event at last exported position. This can happen because if the node is lagging behind in replication, it receives a snapshot and throws away it logs. After receiving the snapshot exporter will be reset which tries to recover the state from the new snapshot. But the replication of the events might be delayed.
It is arguable if we want to keep this check for the leader. But we can safely assume that this node never become the leader if it doesnot have an uptodate log.