feat(raft): do not exponential backoff when sending heartbeats
This exponential back-off unnecessarily delays sending heartbeat to a temporarily disconnected follower. It leads to election timeout in the follower and it starts a leader election unnecessarily. By removing this backoff, leader attempts to send heartbeats every heartbeat interval even if the follower is away. Note:- This is not completely achieved as there are other parameters that influences it. See https://github.com/camunda-cloud/zeebe/issues/7364 for the discussion.