feat(journal): do not reset readers when deleting segments
When reseting the readers transparently, there is always a possibility that these readers are accessed concurrently in another thread. This can lead to issues which are difficult to debug. We remove this automatic reset and expects the readers seek to the new segments when necessary.

In this case, we are compacting the segments. The assumption is that there are no readers which are still reading from the segments to be deleted.