test: create install req handling test class
One test which makes sure that the follower is able to become leader, to make sure that even when log was compacted the new snapshot is readable and he can work with it.
One test which makes sure that the follower after receiving and snapshot and doing some replay can also become leader and complete an instance, ensures that replay still works.