refactor(broker): harmonize interface
The BrokerAdminServiceImpl was a bit inconsistent in that some of its dependencies were given at construction time, while other injected later. In the future, the admin service will be one of the first components to be created, so all dependencies need to be injected later on.
This change also puts more thoughts into the behavior of the class before everything is injected. Ultimately, it would be great if the class could communicate to the caller that certain components are not ready yet. For the time being, however, neutral NoOp implementations are used. In effect, if the admin interface is used before everything is injected, nothing bad will happen.