feat(broker): close zeebedb in ZeebeDbPartitionStep
StateControllerStep is now a startup set. It will be created only once during the startup. Then during role transitions zeebedb will not be closed as statecontroller is not closed during transitions. Since ZeebeDbPartitionStep is responsible for opening the database, it make sense to close it also in this step. This will be more relevant when all steps are split into startup or transition steps.