refactor(broker): remove step timeout
The step timeout was used sparingly and not consistently. If it would be applied consistently, then setting a timeout for all steps is challenging.
The value needs to be long enough for the longest step. But thereby it also gives too much time for short steps.

With the move to K8S environments it will be easier to set a global timeout for the system to get ready.