refactor(engine): improve record creation helper
When we want to create test records we can generally use the
protocol-jackson module. However, in the engine's state module, we
generally use our own implementation of these records which
protocol-jackson does not extend. The engine provides a Records util
class to easily create some records for testing purposes.

For the upcoming tests, I needed the test records to be easily created
and they require both a key, and a processId (+ anything that makes them
valid records when writing/reading them from the state).

This commit adds a new helper method to create valid process records,
and expands the possibilities of others to easily pass in the processId.