feat(transport): use request type to set topic name
The AtomixServerTransport was able to subscribe a request handler to
requests for a partition id. To be able to differentiate between request
types this allows request handlers to be subscribed to requests of a
specific request type for a partition id.

This subscription works through the topicName, which already contained
a hardcoded `command` prefix. By replacing this with a lowercase request
type, the code stays backwards compatible.

Subscriptions happen individually per partition id and request type,
while unsubscribing can be done for a partition in one go.

A default request type is assigned to all ClientRequests while we are
still refactoring and can be removed later.