feat(protocol): add execute query messages
Add new SBE message definitions to be able to request query execution.
These messages contain the bare minimum required to make the request and
respond to it. This helps to avoid using these messages for more than
their intended purpose, which is to support the temporary broker query
api.

With the partitionId, key and valueType we can use the engine's
StateQueryService to query for a bpmnProcessId.