refactor(broker): inject query service on becoming leader
In order to provide access to the query service for the broker query
api, we inject it via the onBecomingLeader method of PartitionLeader.

The upcoming PartitionManager is not yet available to improve how the
broker can gain access to such a service. Like the LogStream injection,
we should change this when it is available. Until then, we can use this
method.