refactor(zeebe-util): replace submit with run
The ConcurrencyControl interface was intended as a reduced interface for the ActorScheduler logic.
The ActorScheduler logic offers run() and submit() methods which do pretty much the same thing,
they schedule a task. Out of those two, run() is more versatile. It can be called from actor threads
and non actor threads. Because of that, run() is preferable over submit().