fix(protocol-test-util): deploy is done when fully distributed
When using the PartitionTestClient to deploy a process, we should wait
until the deployment is fully distributed. Distributed is only written
when it actually distributed the deployment to other partitions.
However, if there is only a single partition, then it does not
distribute it and simply immediately writes FULLY_DISTRIBUTED to the
log. We are certain the process has been distributed to all available
partitions once that record is written. That works for both single as
well as multi partition scenarios.