fix(protocol): remove redundant query response fields
The partitionId, key and valueType properties are not necessary in the
response, because these are already known (they are part of the
request). Matching the response to a specific request is not generally
achieved on the contents of the response, but rather on some other
identifier (like the request id).