refactor(util): allow submitting to an actor from outside
Refactors the ActorControl#submit method to allow submitting tasks from
outside of an actor thread. This was previously only possible via run or
call, but these had confusing semantics: when run was called from
within an actor thread, it would prepend its job (or rather, append it
to the high priority queue). When called from the outside, it was append
its job (i.e. append it to the normal queue). With the introduction of
interfaces to better handle interfacing between actors/non actors, this
becomes very confusing.

This commit fixes submit such that it now be used outside and inside an
actor thread. From within an actor thread, it behaves exactly as before.
From without, it now behaves like run behaves. This way, we have a
method which allows appending tasks to the normal queue in a consistent
way both from within and without an actor thread.

The one downside here is that submitting many jobs means constantly
creating new ActorJob objects, and not reusing the ones from the
reusable task pool. This causes more allocations - but as we're already
using run or call for such cases, which are doing this, I don't think
this is a major worry (at least not introduced in this commit).