fix: write createl metrics in processor
This means we need to create the metrics class on a much higher level, and have to pass it through several layers which is not nice but the simples solution for now. Alternative was to create a singleton but I prefer not to do that, since we have then other issues. OR split it up in several metrics, such that we can have multiple metric classes one only for the created job metric which can be used in the behavior, but this is also more work which Idk whether this is currently necessary.