fix(engine): thread safe closed state
Since query service is deregistered from the query service API via a
listener, and listeners are executed only after the transition is
completed, there was an edge case where the query service might be
accessed in the API while handling a request, but might also be closed
due to a concurrent transition of the partition. Since it relies on the
underlying ZeebeDb, which would also be closed, any access would produce
a segmentation fault.

To prevent this, we introduce a volatile isClosed flag in the service,
which is toggled during the transition before closing the ZeebeDb used.
This ensures that we don't access ZeebeDb once it's closed in a thread
safe way.

While having to check a volatile flag on every call is not performant,
since the QueryService is meant to be removed in the long term in favor
of multi tenancy, and since the service in itself is already fast in
general, it was deemed the simpler solution. Other options included
wrapping the service in an actor, which has higher complexity, but would
be the correct way of doing things for normal features.