fix(broker): fix concurrency access of request handler
Fixes concurrent access of the query service map and request handling
when these occur in different actors/threads, by making the request
handler itself an actor. This obviates the need for a command queue such
as the one present in the CommandAPI.

Also handles the case where the service is closed concurrently by adding
a new partition transition state to build and close the service, to
ensure that no segmentation faults occur due to concurrent transition
and request handling.

Additionally, moves the query API configuration from the network
section to the experimental section.