feat(broker): add new transition steps
All steps from LEADER_STEPS and FOLLOWER_STEPS are already re-implemented as startup steps and transitions steps. In this commit, we add the transition steps to the list in the order that it is expected to be installed.
This steps will be used when the flag for using new transition logic is enabled.