fix: set client request timeout
Elastic search seem to not response on some requests, which caused IT test to fail. The elastic rest client was configured without any request timeout which caused a deadlock. This fix adds a new request timeout config (default 1 second), which allows to fail the request after one second and retry.