feat(gateway): load arbitrary interceptors
Loads arbitrary interceptors as part of the gateway startup, and wraps
the gateway service with them in a chain.

Interceptors can be loaded from external JARs, and will be loaded with
isolated class loaders to avoid dependency collisions.

Each interceptor is decorated with a special internal interceptor which
forwards the intercept call, ensuring that for that call, the thread
context class loader is correctly set.

An executor is provided in the context, accessible via InterceptorUtil,
which lets users use asynchronous code and pipe it back to it to ensure
their call sets the right thread context class loader.