feat(gateway): inject query API in gRPC context
Injects the query API in the context of every call via the outermost
interceptor ContextInjectingInterceptor.

Removes the context executor, as it's complexity we simply don't need.
Pitfalls are well documented, and if we see this is something users
actually run into a lot, we can look into a better solution.