fix(broker): do not set old exported position
When a follower receives exported position from the leader, it is not always the latest exported position. If a new leader is started and it recovers from an older snapshot, the exported position that it sends to the follower might be older than what the follower has already received from the old leader. This can cause to inconsistent state in snapshot, as a previous snapshot might have already compacted the events at the exported position from the new leader. Hence when receiving an exporter position, follower should update its state only if it is larger that what it has already observed.

(cherry picked from commit c087b106d215ec006a0f8e2d512786413bd5c6bd)