refactor(broker): migrate command API step
This is a bit of a departure from the pattern to create components in the constructor of the Broker and then to just start them in the steps.
Reason is that we need to create three objects, but are in the end only interested in the last one.

I also thought about introducing a command API actor, that then internally builds the other components. And I thought about extending the
commandApiServiceImpl, which already is an actor. But none of those options felt fitting.

Therefore, all three components are now created and wired in the startup step. In the startup context all three are visible and known
by their implementation class. In the broker context, only the command API is visible and only by its interface.