refactor(broker): register additional listeners up front
Removes a method from the Broker which was used by tests to register additional partition listeners.
The issue with this method was that it was not clearly tied to the lifecycle of the Broker.
To make it clearer that listeners need to be known when the Broker is started, this
commit adds a constructor parameter.

This is also a preparatory steps for migrating the bootstrap step that creates the partition manager
and the partitions. This will make the code related to partition listeners easier.