feat(snapshot): enable snapshot store to copy snapshot to a target directory
It can happen that while a snapshot is being deleted by the snapshot store, state controller is trying to copy the snapshot to the runtime. In order to sycnhronize access to the snapshots, all operations that can access a snapshot should go through the snapshot store. In this commit we add "copySnapshot" method to the snapshot store.