fix(broker): skip transitions if it is behind the latest term
This adds checks whether the term associated with the current task is the latest term.
If it is not, the current task logs an information that it is obsolete and completes
the transition future associated with the task.

This check needs to be done at three points in time:
- before the current transition is cancelled
- before the current transition will be cleaned up
- before the new transition is created

As soon as the new transition is created, we are safe again.

The second change is to make the transition cleanup idempotent. Even with the checks
in place it could be that one transition already started the cleanup. And while the
cleanup is happening a new transition comes in. In this case the new transition will trigger
a second cleanup. This is now possible with the change