fix: execute transitions in order
In order to resolve our current concurrency issues (executing multiple
transitions alternately) we decided to go back to enqueuing
transitions. This means we start one transition and if a new transition
comes in we register on the future to enqueue the next transition. We
cancel the previous transition to make the transition process faster.

This means if we have a running transition and two new transitions come
in then the first one is canceled, the second one is skipped (but the
resources of the first one are closed) and the third one is installed.