refactor: make actor context umodifiable
Add new method which should be overwritten by subclasses in order to add more information on which context the actor is currently running. This method is only on the first time, when the context is not yet filled.

(cherry picked from commit 2091b0f0ea771ffb92e89603241bdc3fe387a925)