refactor(broker): remove statecontroller step
StateControllerPartitionStep only creates the statecontroller instance. There is nothing else to do. So it is enough to create it statically once when the context is created. A step is unnecessary. It is also not necessary to close the stateController as ZeebeDb step is responsible for opening and closing the db.