fix(broker): run stateController operations on SnaphostStore actor
It can happen sometimes that the an attempt to take a snapshot is done after ZeebeDb is closed. This happens because ZeebeDb is closed by ZeebeDbPartitionTransitionStep which is running on ZeebePartition actor, while the snapshot is taken on the SnapshotStore actor. As a result, it can happen that the job to take snapshot is still queued on the snapshot store actor even after ZeebeDb is closed. Inorder to prevent, this we now ensure that StateController methods are run on the SnapshotStore actor. This is done by setting the concurrency control of StateControllerImpl to the SnapshotStore actor. In this way all access to the database and snapshot is synchronized via the same actor.

* To enable state controller to run on an actor, we changed all methods in StateController interface to return future. All usages of StateController is also changed to use the new interface.
We also removed "openDb" from the interface. Instead StateController#recover should open the database after recovery.
* All public methods on StateControllerImpl executed using provided concurrencyControl
* Updated StateControllerImplTest to use the new interface
* Set the concurrencyControl of state controller to use the same actor as snapshotStore.
* Replace call with submit in snapshotStore#addPendingSnapshot. This method is called from the same actor. So it should not be "call". "submi"t appends the task to the queue when called from the same actor or outside. Hence in this context it is better to use submit. "run" prepends when called from the same actor, and appends when called from outside. This method can be called from both with in and outside snapshot store actor. Hence using "submit" gives a consistent behaviour.