fix(broker): fix flaky test
The reason for the test failures are unknown. However, this test failed repeatedly and always with this root exception:

Caused by: java.lang.ClassCastException: class org.mockito.codegen.Set$MockitoMock$203028413 cannot be cast to class io.atomix.cluster.Member (org.mockito.codegen.Set$MockitoMock$203028413 and io.atomix.cluster.Member are in unnamed module of loader 'app')
	at io.atomix.cluster.ClusterMembershipService$MockitoMock$685621782.getLocalMember(Unknown Source)
	at io.camunda.zeebe.broker.partitioning.PartitionFactory.constructPartitions(PartitionFactory.java:164)
	at io.camunda.zeebe.broker.partitioning.PartitionManagerImpl.lambda$start$3(PartitionManagerImpl.java:151)
	at java.base/java.util.concurrent.CompletableFuture.uniApplyNow(CompletableFuture.java:680)
	... 11 more

It is unclear why this can happen. It is also unclear why this does not happen all the time.

Be that as it may, this commit provides a real implementation for the member object.