fix(broker): always trigger transitions
PartitionTransitionSteps are already optimizing the transitions by deciding whether to re-install the services during the transition. Hence this additional check here is unnecessary. Besides this check did not consider transitioning from inactive to a follower role, which made some of the new tests fail.
I also think it is error prone to keep the raftRole in multiple places. So it is better to use this only for logging, and not for any decisions.