refactor: split processing panel
Replace processing panel with two panels, which one shows the processing (and writtes) and one which shows the exporting. This helps to spot when exporting or processing has stopped. In general as part of separate concerns. The panels got two new limits, min = 0 and max = 5000, normally an partition can process/exporter ~ 2.5k events/records. Sometimes the writting spikes, which breaks totally the view, the limits ensure to make the metrics still visible.