feat(atomix/cluster): allow secure inter-node communication over TLS
Introduces security configuration similar to the gateway's for
inter-node communication. Users can now configure a certificate chain
and private key which is loaded by the NettyMessagingService in order to
encrypt communication between all cluster nodes. TLS version is fixed at
1.3, and uses BoringSSL via the tcnative artifact provided by Netty.

This solution assumes all nodes share the same certificates, as we load
the certificate chain as part of the trust manager for client
connections, and there's no real way to specify a different trust
manager. Assuming they share the same intermediate chain, it could be
possible to have different final certificates, but this is not tested.