refactor(dist): configure monitoring via Spring
Configures the monitoring server via Spring properties and ignores the
broker/gateway's monitoring configuration.

The properties used are the server.address and server.port, and not the
management.server.address or management.server.port. This is because the
latter are used only when you want to differentiate the normal MVC bind
address from the management server's. In our case, we only use the
embedded server as the management server, so there's no need to
distinguish them. It's also not possible to use only the management.*
properties and disable the server.* properties.