fix(gateway): instantiate the monitoring interceptor once
Only instantiate the monitoring interceptor as a temporary fix for
metrics being registered multiple times.

Ultimately, the proper solution is to allow injecting a collector
registry at runtime and binding the metrics to it. This will allow us to
unit test metrics, something which isn't possible, and will ensure we
can parallelize tests without them interfering with each other. It will
also be beneficial when we move inter-cluster communication to use gRPC,
since we will have multiple interceptors in different places.