ci(.github): fix dependabot commit types
Our CONTRIBUTING.md reserves `deps` as a commit type for changes to
external dependencies and even specifically mentions Dependabot. Let's
use the these. This also allows to differentiate between different types
of dependencies (i.e. maven for maven dependencies, go for the go client
and .github for the github actions).

Note that 'clients/go' was not possible as scope in the prefix, because
of a maximum of 15 chars for the prefix.