fix(engine): modify wait condition
This commit modifies the wait condition to circumvent the bug described in #7995.

As a result this test is no longer flaky. Before, the test failed 1 in ~150 times. With this change it did not fail once in 2000 iterations.

(cherry picked from commit 88c87221f65bbe72e87d45cb867bf211f6ad5a50)