refactor(atomix): improve error handling
This adds convenience methods when setting the cluster's messaging
security which also take in strings additionally to File instances. This
is to improve the error handling when there is no configured value, and
return an appropriate error message.

Additionally improves error messages when there is nothing configured
for either the certificate chain or the private key, or when one of them
is null, missing, or not readable.