feat(broker): allow secure inter-cluster messaging
Introduces a new network configuration property, security, to configure
secure inter-cluster messaging. This allows users to enable TLS
encryption using a certificate chain and a private key for all
inter-cluster communication: the command API and the internal API.

Validation of the configuration is performed in the SystemContext and
the broker will fail to start if security is enabled but either the
certificate or private key is null, or not a readable file.