fix(logstreams): complete future when any error during opening an appender
Not all exceptions occuring during opening an appender was handled properly. In such cases the caller (StreamProcessor) waiting for the writer will wait for ever because the future is not completed. This commit adds a catch all exception to the openAppender which completes the futuer exceptionally.

(cherry picked from commit 56376ed2eeeacb12d6448746347b1cc4352b3dab)