refactor(broker): always set stream processor in context
With this commit the stream processor is always set in the transition context, regardless of whether it could
be started successfully or not. The main benefit is that the reference to the newly created stream processor
is always stored in the context, and so it can be found there and closed.

Previously, if the opening of the stream processor failed, it was in an undefined state - potentially running
and potentially registered in ActorScheduler which holds a permanent reference to it.