ci: only ratched when formatting
We don't want to use ratched on the build pipeline, because it requires
us to have origin/develop fetched. The check does not require
ratchetting and to be honest is better without ratchetting. By moving
the ratchet setting into the execution of the apply we make sure it is
only used locally.