fix(raft): use a temporary reader for finding last zeebe entry in LeaderRole
In preparation to removing the long living reader in RaftContext, we want to remove all usages of it.

In this commit, the LeaderRole uses a temporary reader to find the last zeebe entry in LeaderRole. This operation is only done once, when the LeaderRole is started. Hence creating a new reader here is acceptable.

(cherry picked from commit 8545d57ae3a15ae08b1573bfe11d033f331a2c89)