fix(raft): use a temporary reader in PassiveRole for checking consistency
In preparation to removing the long living reader in RaftContext, we want to remove all usages of it.

In this commit, the PassiveRole uses a temporary reader when checking the pre-conditions for accepting an AppendRequest. The reader is only used during special conditions when the previous entry in the log does not match with the expected previous entry. This case happens only rarely. So creating a new reader everytime does not impact performance.

(cherry picked from commit b436deee4f6296662cb536954d71725f74ebd13b)