feat(engine): add static value is bracketed list check
Adds a function to check whether a static value is a bracketed list of
comma-separated values.

Originally, I considered to implement this using a regular expression,
but complexity and a potential stackoverflow exception (highlighted by
sonar) urged us to move away from that solution.

An other idea is build a parser, which can directly be used in the
transformer that needs to be written later. For now, we chose an easy
route and can always change this implementation later if necessary.