feat(engine): validate assignment definitions
Adds the ZeebeAssignmentDefinition validations. Specifically, it
verifies assignee and candidate groups.

Both are optional, and can be static values and expressions.
For candidate groups special rules apply when it is a static value.

No validation is added to make sure either one of assignee or
candidateGroups is provided, because such a restrictive check seems not
so useful and no tools are available at this time on the validators to
suport such a check.