build(monitor): fix processing metrics
Fixes the processing metrics visualization. These would sum the various
positions. That was not an issue to aggregate values when there was only
one leader in the given time interval, but with multiple, the positions
would get summed, resulting in much larger positions than exist on the
log. Similarly, the processing queue ended up being much larger than the
real processing queue was.

The fix was to use `max` as an aggregating function instead of `sum`.
This works when dealing with absolute values like positions, where we
only care about the last position (or about subtracting two positions).