test: support writing record batch with source record position
When we want to write a btach of records in a test the source record
position can be important. Specifically, when the engine is restarted or
replay is initiated and some commands with their events have been
written in the batch, then the commands are only processed if any events
referred to it with their source record position.

So if you intent to write a command and some events where those events
are the result of virtually processing the command, then you should
write those events with the correct source record position.