fix(engine): candidateGroups dont have to be bracketed
I'm unsure where I got in my head that candidate groups as static value
should be a bracketed list of csv, but that was wrong. What we actually
wanted was for static candidateGroups to be defined simply as csv
(without the brackets). This commit applies the changes to fix this
formatting validation