build(benchmarks): adds release script for images
Adds a release script for the benchmark worker/starter images which will
checkout the version you need in a worktree (so isolated), build the
images for that version, and optionally push them. To use, you would go
to the benchmarks/project folder and run `./release.sh -p VERSION`, e.g.
`./release.sh -p 1.2.4`.