build(ci): disable invalid incremental compilation
It seems incremental compilation causes some modules to be recompiled
every time. This can lead to race conditions in CI when we have multiple
stages in parallel which may be recompiling.

See the following issue related to this:
https://issues.apache.org/jira/browse/MCOMPILER-209

The flag should be set to false, which delegates the incremental
compilation duty to the compiler itself (which in most cases takes care
of it properly). This reduces race conditions when parallel stages may
be compiling the same thing, and speeds up CI a little bit.