test(clients): fixes flakiness in a time sensitive test
The test was expecting a timeout after 5 milliseconds, which would
trigger because the server takes 10 milliseconds. This is a very small
time range which can easily be stalled by GC pauses, for example,
resulting in the test sometimes failing as no timeout occurred.

There's no reason not to increase the server's delay to ensure we always
timeout.