refactor(engine): replace tuple with explicit type
When evaluating the job expressions, each of the expressions may fail
and thus each of these is captured by either a failure or some result
value. However, we'll need to evaluate more of these in the future, so
the zeebe tuple (which only fits 2 values) is no longer suitable.

Instead we can replace the tuple with a record-like class and provide it
with specialized setters that return the instance. This allows us to
cleanup the result collection logic.

We start the flatmapping sequence with a right of an empty JobProperties
object, so each of the actual evaluations can be placed under eachother
in flatMap calls. This keeps everything nicely readable