Make ZeebeObjectMapper use a preconfigured one
ATM, it is not possible to configure the ZeebeObjectMapper with a provided and preconfigured ObjectMapper. This makes it impossible to configure the used ObjectMapper (for example to use JSR310 java.time Library). This change will not affect the other implementations in first place.

It just adds the possibility to use my own ObjectMapper containing all required features for my worker.