refactor(atomix): partition metadata tracks primary
Additionally to keeping track of the target priority, the partition
metadata now tracks the primary member. It should be the member with
the highest priority, except in the case of a fixed distribution where
all more than one member has the target priority, in which case no
primary is provided.