fix(engine): support spaces in static csv expressions
Consider 'human, elves' as candidateGroups entry. Spaces should be
allowed in these values and should be trimmed off from each value when
used.

So, additional test cases are added that highlighted the problem that
the expression ' , ' was considered a valid CSV, even though it does not
contain any values.

This was because the splitted values were only checked on being empty,
not blank.