refactor: parse instead of validate expression
By parsing, instead of validating we can make the stream of values
directly available to the transformer as well, reusing the same code.

The parse method returns an either, to make clear that parsing can fail,
although this is not expected to happen in the transformer (since the
value is static and cannot change). This does mean, we can only use this
method for static expressions.