feat(engine): transform task assignment
At this time only user tasks can have task assignments. If they are set,
we need to transform the assignee and the candidate groups to parsed
expressions so these can be evaluated at activation of the user task.

Static values are always CSV formatted (checked by the validators), but
need to be transformed to an expression that doesn't evaluate to the CSV
but rather to a list of strings.