fix(transport): don't reuse buffer for request data
Depending on the implementation of the request handler, reusing buffers
is unsafe. For example, the Query API has an asynchronous request
handler which means that modifying the buffer in `handleAtomixRequest`
is a data race.

Constructing an UnsafeBuffer is relatively cheap, so we can just do it
for every request to be on the safe side. Alternatively we could have
changed the interface for `RequestHandler` to accept the raw byte array
instead of a buffer.

(cherry picked from commit 768cc363f8cff5ea484c436b204d773f31dd2f4e)