fix(broker): ensure readers and writers are always reset
Previously resetting only happened after handling the request.
When a request was not read successfully, the state of the error
response writer could leak from one request to the next.

Instead of trying to add the reset logic at every return point of
`handleRequest` we can simply reset the readers and writers at the start
of every request.