feat(engine): eval candidate groups expression
Evaluates the candidate groups expression when the job is being created
(for an activating task). Like the assignee expression, this evaluation
can fail and result in an incident.

The expression is evaluated as a list of strings, for which we now use
the MsgPackConverter::convertToJson method which results in each string
element to be wrapped by quotes (as a string literal), because that's
how it looks like from a json perspective. Therefore, we need to unwrap
the quotes from the string before collecting it into a list.

The list can then be transformed into a string-representation.