fix(journal): prevent concurrent compaction and opening reader
Compaction and opening a new reader can happen in parallel, because compaction is scheduled on the raft thread while a transition happens in parallel on the zeebe threads which can open a new reader concurrently. This will cause issues because the segment accessed by the new reader might be already deleted by the parallel `deleteUntil` operation. To prevent this, we now also make `deleteUntil` thread-safe.