test(engine): verify evaluateArrayOfStingsExpression
Sadly, we cannot yet test this method to its fullest, because at this
time we do not have a way to check that the elements in the evaluated
list are strings. For now, we've decided to simply allow expressions
that evaluate to a list of any and transform them to list of strings
using the convertToJson method.