fix(broker): fix NPE
The NPE could occur because it is possible to interrupt the startup procedure. When the startup procedure is interrupted, it finishes executing the current startup step, then aborts startup.
After startup is aborted, the shutdown commences, which - among others - sets the healthCheckService to null.

This led to a race condition where the healthCheckService was set to null before we could call it to set the broker as started.

Switching the order of statements fixes this race condition.