fix: reset last written position
Previously the last written position was not reset on starting the processing state machine, which could cause weird metric exports, because the gap between lastWritten and newWritten position was to high (over million and more). The replay state machine now returns all positions and initialized the processing state machine with it.