refactor: exit on unrecoverable errors
Here we are adding the concept of a FatalErrorHandler that ensures that
we handle unrecoverable errors consistently.
The main implementation is VirtualMachineErrorHandler which exits with a
custom status code when the JVM reports VirtualMachineErrors.
We are using the FatalErrorHandler in a couple of different places:
* actors and actor threads
* atomix threads
* as a default uncaught exception handler for all threads

There are still places where we are swallowing errors, most notably
netty's OutOfDirectMemoryError but this patch should already improve
the current situation.