fix(broker): ensure ongoing transition are not stuck while closing ZeebePartition
When `super.closeAsync` is called, it discards jobs from the Actor's queue. This resulted in an ongoing transition to get stuck, because the jobs related to it were discarded. Transition to inactive triggered by the close would then never completes, because it is waiting for the ongoing transition to complete. Thus ZeebePartition cannot be closed.
In this commit, we fix this behavior by ensuring that transitionToActive is called before actor close is requested. We also ensure that no new transitions can be enqueued, once the closing is started. This is important because we don't want new transition after it has all steps are closed by transitioning to inactive.

(cherry picked from commit 3a7cf18494accc2d762d121ccefd260974639591)