test(raft): enable requests to timeout in randomized raft test
Raft expects the messaging layer to timeout the requests. In our mock implementation of the messaging layer, we did not implement the timeout correctly. As a result the requests are pending in raft. This will break some of the liveness guarantees. Ideally raft should timeout requests itself. But since that is a bigger change, we will just fix the test setup to mimic the actual implementation.

(cherry picked from commit 80333ff4120620298823e58961472474b7f9dcb5)